refactor : converter 메서드를 dto로 이동
